Split widget: signal list
